pyproject.toml
[tool.poetry]
name = "zeron"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
pillow = "^10.4.0"
black = "^24.4.2"
django-ckeditor = "^6.7.1"
gunicorn = "^22.0.0"
django-crontab = "^0.7.1"
django-allauth = "^0.63.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
